third_party/blink/web_tests/platform/generic/media/media-fragments/TC0084-expected.txt


Title: TC0084
Fragment: 't=0:60:00.000'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
